refactor(approval-gate): extract sweeper.rs from lib.rs
Seventh step of breaking lib.rs into focused modules. This commit
lifts the periodic timeout sweeper and the small iii-stream helpers
it owns into their own module.

Moved (lib.rs -> sweeper.rs):
- spawn_timeout_sweeper()
- timeout_resolved_event()
- write_event()
- write_hook_reply()
- uuid_like()

All five stay pub(crate) — only the register() wiring in lib.rs
and the resolve flow need them. Pure move, no behavior change.
141 approval-gate tests pass.
